improve things
